requirements.txt
Django==5.1.*
djangorestframework==3.15.*
psycopg[binary]==3.2.*
drf-spectacular
